Draft 2 of mixnet maths. Some cleanup, move extra to test.
