+++ src/main/java/org/cryptobiotic/gmp/RuntimeHelper.java
@@ -39,7 +39,8 @@ final class RuntimeHelper {
             (size, align) -> Arena.ofAuto().allocate(size, align);
 
     static {
-        System.load("/usr/local/lib/libgmp.so");
+        // System.load("/usr/local/lib/libgmp.so");
+        // System.loadLibrary("egkgmp");
         System.load("/usr/local/lib/libegkgmp.so");
         SymbolLookup loaderLookup = SymbolLookup.loaderLookup();
         SYMBOL_LOOKUP = name -> loaderLookup.find(name).or(() -> LINKER.defaultLookup().find(name));
