+++ src/test/kotlin/org/cryptobiotic/gmp/EgkGmpMpz.kt
