+++ src/test/kotlin/org/cryptobiotic/maths/ProdColumnPowJava.kt
