+++ src/test/kotlin/org/cryptobiotic/maths/ProdColumnPowTab.kt
@@ -8,6 +8,7 @@ import kotlinx.coroutines.channels.ReceiveChannel
 import kotlinx.coroutines.channels.produce
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
+import org.cryptobiotic.prodPow.VmnProdPowW
 import kotlin.math.max
 import kotlin.math.min
